Chnaged node_reset from int to float. Implemented a range of values to account for drifting values.

+++ CyTracker/Feather_LORA/LoRa_Firmware/Mission_Control/Data.h
@@ -33,7 +33,7 @@ class DATA
     // Set High for # (defined below) seconds on startup to tell other nodes in the system
     // that an unexpected power cycle has occurred and they need to clear 
     // their locally held variables.
-    int node_reset = 0;
+    float node_reset = 0.0;
     // Number of seconds after startup that the reset_node will be held high.
     unsigned long startup_timer = 0;
     // Turns true after # seconds of running. Used to tell if the system started
